feat: 부동산, 주급 입금 주기 설정 및 수정

+++ Assets/02. Scripts/Outing/Office/OfficePanelController.cs
@@ -198,7 +198,7 @@ public class OfficePanelController : MonoBehaviour {
     void RefreshSalaryInfo() {
         if (OfficeManager.Instance == null || salaryInfoLabel == null || nextSalaryLabel == null) return;
         OfficeManager office = OfficeManager.Instance;
-        int salary = office.currentMonthlySalary;
+        long salary = office.currentMonthlySalary;
         float cycleDays = office.GetSalaryCycleDays();
         float remainDaysRaw = office.GetDaysUntilNextSalary();
         int remainDaysInt = Mathf.CeilToInt(remainDaysRaw);
